git-svn-id: https://svn.kuali.org/repos/student/trunk@16654 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-lum/ks-lum-impl/src/main/java/org/kuali/student/lum/service/assembler/LumServiceMethodInvoker.java
@@ -4,11 +4,12 @@ import java.util.List;
 
 import org.apache.log4j.Logger;
 import org.kuali.student.core.assembly.BaseDTOAssemblyNode;
-import org.kuali.student.core.assembly.BusinessServiceMethodInvoker;
 import org.kuali.student.core.assembly.BaseDTOAssemblyNode.NodeOperation;
+import org.kuali.student.core.assembly.BusinessServiceMethodInvoker;
 import org.kuali.student.core.assembly.data.AssemblyException;
 import org.kuali.student.core.atp.service.AtpService;
 import org.kuali.student.core.exceptions.AlreadyExistsException;
+import org.kuali.student.core.exceptions.CircularReferenceException;
 import org.kuali.student.core.exceptions.CircularRelationshipException;
 import org.kuali.student.core.exceptions.DataValidationErrorException;
 import org.kuali.student.core.exceptions.DependentObjectsExistException;
@@ -20,58 +21,106 @@ import org.kuali.student.core.exceptions.PermissionDeniedException;
 import org.kuali.student.core.exceptions.UnsupportedActionException;
 import org.kuali.student.core.exceptions.VersionMismatchException;
 import org.kuali.student.core.organization.service.OrganizationService;
+import org.kuali.student.core.statement.dto.RefStatementRelationInfo;
+import org.kuali.student.core.statement.dto.ReqComponentInfo;
+import org.kuali.student.core.statement.dto.StatementInfo;
+import org.kuali.student.core.statement.dto.StatementTreeViewInfo;
 import org.kuali.student.core.statement.service.StatementService;
 import org.kuali.student.lum.course.service.assembler.LoCategoryRelationInfo;
 import org.kuali.student.lum.lo.dto.LoInfo;
 import org.kuali.student.lum.lo.dto.LoLoRelationInfo;
 import org.kuali.student.lum.lo.service.LearningObjectiveService;
+import org.kuali.student.lum.lrc.dto.ResultComponentInfo;
+import org.kuali.student.lum.lrc.service.LrcService;
 import org.kuali.student.lum.lu.dto.CluCluRelationInfo;
 import org.kuali.student.lum.lu.dto.CluInfo;
 import org.kuali.student.lum.lu.dto.CluLoRelationInfo;
 import org.kuali.student.lum.lu.dto.CluResultInfo;
 import org.kuali.student.lum.lu.service.LuService;
 
-public abstract class LumServiceMethodInvoker implements BusinessServiceMethodInvoker {
+public class LumServiceMethodInvoker implements BusinessServiceMethodInvoker {
 	final Logger LOG = Logger.getLogger(LumServiceMethodInvoker.class);
 	private LuService luService;
 	private StatementService statementService;
 	private LearningObjectiveService loService;
 	private OrganizationService orgService;
 	private AtpService atpService;
-	
+	private LrcService lrcService;
+
 	@SuppressWarnings("unchecked")
-	@Override
-	public void invokeServiceCalls(BaseDTOAssemblyNode results)
+	public final void invokeServiceCalls(BaseDTOAssemblyNode results)
 			throws AlreadyExistsException, DataValidationErrorException,
 			DoesNotExistException, InvalidParameterException,
 			MissingParameterException, OperationFailedException,
 			PermissionDeniedException, VersionMismatchException,
 			DependentObjectsExistException, CircularRelationshipException,
-			AssemblyException, UnsupportedActionException {
+			AssemblyException, UnsupportedActionException, UnsupportedOperationException, CircularReferenceException {
 
-	    // For Delete operation process the tree from bottom up 
+	    // For Delete operation process the tree from bottom up
 	    if(NodeOperation.DELETE == results.getOperation()) {
             for(BaseDTOAssemblyNode childNode: (List<BaseDTOAssemblyNode>) results.getChildNodes()){
                 invokeServiceCalls(childNode);
             }
 	    }
-	    
-	    Object nodeData = results.getNodeData();
+
+	    invokeServiceCallOnResult(results);
+
+		// For create/update process the child nodes from top to bottom
+		if(NodeOperation.DELETE != results.getOperation()) {
+		    for(BaseDTOAssemblyNode childNode: (List<BaseDTOAssemblyNode>) results.getChildNodes()){
+		        invokeServiceCalls(childNode);
+		    }
+		}
+	}
+
+	/**
+	 * @param results
+	 * @throws AlreadyExistsException
+	 * @throws DataValidationErrorException
+	 * @throws DoesNotExistException
+	 * @throws InvalidParameterException
+	 * @throws MissingParameterException
+	 * @throws OperationFailedException
+	 * @throws PermissionDeniedException
+	 * @throws AssemblyException
+	 * @throws VersionMismatchException
+	 * @throws DependentObjectsExistException
+	 * @throws CircularRelationshipException
+	 * @throws UnsupportedActionException
+	 * @throws UnsupportedOperationException
+	 * @throws CircularReferenceException
+	 */
+	protected void invokeServiceCallOnResult(BaseDTOAssemblyNode results)
+			throws AlreadyExistsException, DataValidationErrorException,
+			DoesNotExistException, InvalidParameterException,
+			MissingParameterException, OperationFailedException,
+			PermissionDeniedException, AssemblyException,
+			VersionMismatchException, DependentObjectsExistException,
+			CircularRelationshipException, UnsupportedActionException,
+			UnsupportedOperationException, CircularReferenceException {
+		Object nodeData = results.getNodeData();
+		if (nodeData == null) {
+			return;
+		}
 		if(nodeData instanceof CluInfo){
 			CluInfo clu = (CluInfo) nodeData;
 			switch(results.getOperation()){
 			case CREATE:
 				CluInfo newClu = luService.createClu(clu.getType(), clu);
-				results.getAssembler().assemble(newClu, results.getBusinessDTORef(), true);
+				if(results.getAssembler() != null) {
+					results.getAssembler().assemble(newClu, results.getBusinessDTORef(), true);
+				}
 				break;
 			case UPDATE:
 				CluInfo updatedClu = luService.updateClu(clu.getId(), clu);
-				results.getAssembler().assemble(updatedClu, results.getBusinessDTORef(), true);
+				if(results.getAssembler() != null) {
+					results.getAssembler().assemble(updatedClu, results.getBusinessDTORef(), true);
+				}
 				break;
 			case DELETE:
 				luService.deleteClu(clu.getId());
 				break;
-			}						
+			}
 		}else if(nodeData instanceof CluCluRelationInfo){
 			CluCluRelationInfo  relation = (CluCluRelationInfo) nodeData;
 			switch(results.getOperation()){
@@ -80,10 +129,10 @@ public abstract class LumServiceMethodInvoker implements BusinessServiceMethodIn
 				// Update the businessDTO if one exists for the cluclurelation (for e.g. CourseJointInfo)
 				if(null != results.getBusinessDTORef()) {
 					results.getAssembler().assemble(newCluRel, results.getBusinessDTORef(), true);
-				}				
+				}
 				break;
 			case UPDATE:
-				CluCluRelationInfo updatedCluRel = luService.updateCluCluRelation(relation.getId(), relation);				
+				CluCluRelationInfo updatedCluRel = luService.updateCluCluRelation(relation.getId(), relation);
 				// Update the businessDTO if one exists for the cluclurelation (for e.g. CourseJointInfo)
 				if(null != results.getBusinessDTORef()) {
 					results.getAssembler().assemble(updatedCluRel, results.getBusinessDTORef(), true);
@@ -92,7 +141,7 @@ public abstract class LumServiceMethodInvoker implements BusinessServiceMethodIn
 			case DELETE:
 				luService.deleteCluCluRelation(relation.getId());
 				break;
-			}			
+			}
 		}else if(nodeData instanceof CluResultInfo){
 			CluResultInfo cluResult = (CluResultInfo) nodeData;
 			switch(results.getOperation()){
@@ -170,16 +219,105 @@ public abstract class LumServiceMethodInvoker implements BusinessServiceMethodIn
 				luService.deleteCluLoRelation(cluLoRelation.getId());
 				break;
 			}
-		}
-		
-		// For create/update process the child nodes from top to bottom
-		if(NodeOperation.DELETE != results.getOperation()) {
-		    for(BaseDTOAssemblyNode childNode: (List<BaseDTOAssemblyNode>) results.getChildNodes()){
-		        invokeServiceCalls(childNode);
-		    }
+		}else if(nodeData instanceof ResultComponentInfo){
+			ResultComponentInfo resultComponent = (ResultComponentInfo) nodeData;
+			switch(results.getOperation()){
+			case CREATE:
+				ResultComponentInfo createdResultComponent = lrcService.createResultComponent(resultComponent.getType(), resultComponent);
+				//Copy the created back to the reference Should there be an assembler for this?
+				if(results.getBusinessDTORef()!=null&& results.getBusinessDTORef() instanceof ResultComponentInfo){
+					ResultComponentInfo resultComponentToUpdate = (ResultComponentInfo) results.getBusinessDTORef();
+					resultComponentToUpdate.setId(createdResultComponent.getId());
+					resultComponentToUpdate.setType(createdResultComponent.getType());
+					resultComponentToUpdate.setDesc(createdResultComponent.getDesc());
+					resultComponentToUpdate.setEffectiveDate(createdResultComponent.getEffectiveDate());
+					resultComponentToUpdate.setExpirationDate(createdResultComponent.getExpirationDate());
+					resultComponentToUpdate.setMetaInfo(createdResultComponent.getMetaInfo());
+					resultComponentToUpdate.setName(createdResultComponent.getName());
+					resultComponentToUpdate.setResultValues(createdResultComponent.getResultValues());
+					resultComponentToUpdate.setState(createdResultComponent.getState());
+				}
+				break;
+			case UPDATE:
+				lrcService.updateResultComponent(resultComponent.getId(), resultComponent);
+				break;
+			case DELETE:
+				lrcService.deleteResultComponent(resultComponent.getId());
+				break;
+			}
+		} else if(nodeData instanceof RefStatementRelationInfo){
+			RefStatementRelationInfo relation = (RefStatementRelationInfo) nodeData;
+			switch(results.getOperation()){
+			case CREATE:
+				RefStatementRelationInfo created = statementService.createRefStatementRelation(relation);
+				relation.setMetaInfo(created.getMetaInfo());
+				break;
+			case UPDATE:
+				RefStatementRelationInfo updated = statementService.updateRefStatementRelation(relation.getId(), relation);
+				relation.setMetaInfo(updated.getMetaInfo());
+				break;
+			case DELETE:
+				statementService.deleteRefStatementRelation(relation.getId());
+				break;
+			}
+		} else if(nodeData instanceof StatementInfo){
+			StatementInfo statement = (StatementInfo) nodeData;
+			switch(results.getOperation()){
+			case CREATE:
+				StatementInfo created = statementService.createStatement(statement.getType(), statement);
+				if(results.getAssembler() != null && results.getBusinessDTORef() != null) {
+					results.getAssembler().assemble(created, results.getBusinessDTORef(), true);
+				}
+				break;
+			case UPDATE:
+				StatementInfo updated = statementService.updateStatement(statement.getId(), statement);
+				if(results.getAssembler() != null && results.getBusinessDTORef() != null) {
+					results.getAssembler().assemble(updated, results.getBusinessDTORef(), true);
+				}
+				break;
+			case DELETE:
+				statementService.deleteStatement(statement.getId());
+				break;
+			}
+		} else if(nodeData instanceof ReqComponentInfo){
+			ReqComponentInfo reqComp = (ReqComponentInfo) nodeData;
+			switch(results.getOperation()){
+			case CREATE:
+				ReqComponentInfo created = statementService.createReqComponent(reqComp.getType(), reqComp);
+				reqComp.setMetaInfo(created.getMetaInfo());
+				break;
+			case UPDATE:
+				ReqComponentInfo updated = statementService.updateReqComponent(reqComp.getId(), reqComp);
+				reqComp.setMetaInfo(updated.getMetaInfo());
+				break;
+			case DELETE:
+				statementService.deleteReqComponent(reqComp.getId());
+				break;
+			}
+		}else if(nodeData instanceof StatementTreeViewInfo){
+			StatementTreeViewInfo treeView = (StatementTreeViewInfo) nodeData;
+			switch(results.getOperation()){
+			case CREATE:
+				StatementTreeViewInfo created = statementService.createStatementTreeView(treeView);
+				if(results.getAssembler() != null && results.getBusinessDTORef() != null) {
+					results.getAssembler().assemble(created, results.getBusinessDTORef(), true);
+				}
+				break;
+			case UPDATE:
+				StatementTreeViewInfo updated = statementService.updateStatementTreeView(treeView.getId(), treeView);
+				if(results.getAssembler() != null && results.getBusinessDTORef() != null) {
+					results.getAssembler().assemble(updated, results.getBusinessDTORef(), true);
+				}
+				break;
+			case DELETE:
+				statementService.deleteStatementTreeView(treeView.getId());
+				break;
+			}
+		}else{
+			throw new UnsupportedActionException("This service invoker does not know how to handle nodeData for "+nodeData.getClass().getName());
 		}
 	}
-	
+
 	public LuService getLuService() {
 		return luService;
 	}
@@ -218,6 +356,10 @@ public abstract class LumServiceMethodInvoker implements BusinessServiceMethodIn
 
 	public void setAtpService(AtpService atpService) {
 		this.atpService = atpService;
+	}
+
+	public void setLrcService(LrcService lrcService) {
+		this.lrcService = lrcService;
 	}
 
 }
